separate_frame_orgin/requirements.txt
tornado==4.3
Pillow==7.2.0
pip==20.1.1
setuptools==47.3.1
Image==1.5.32
asgiref==3.2.10
django==3.0.8
sqlparse==0.3.1